detect if a particular kind of object is in a photo

+++ Assets/PhotoScript.cs
@@ -3,15 +3,19 @@ using System.Collections;
 using System.Collections.Generic;
 using System.ComponentModel;
 using UnityEngine;
+using static UnityEditor.Experimental.GraphView.GraphView;
 
 public class Photo
 {
     public Photo(RenderTexture texture)
     {
         this.texture = texture;
+        this.isRedCubeIn = false;
     }
 
     public RenderTexture texture { get; }
+
+    public bool isRedCubeIn { get; set; }
 
     // Eventually add flags for what photos contain
 }
@@ -22,6 +26,8 @@ public class PhotoScript : MonoBehaviour
     public const bool DEBUG_MODE = true;
     public const float DEBUG_SPEED = 0.01f;
     public Vector3 DEBUG_CENTER;
+    // How much of the area the object has to be of the photo to be a "good shot"
+    public const float REQ_AREA = 0.02f;
 
     public GameObject planePrefab;
     private GameObject[] photosPlanes;
@@ -32,13 +38,38 @@ public class PhotoScript : MonoBehaviour
     // Inactive photos are a stack we can pull from and add to the list
     private Stack<Photo> inactivePhotos;
 
+    // We use this RenderTexture to detect if certain objects are in the photo
+    private RenderTexture inspectTexture;
+
+    // We use this RenderTexture to blit to so we don't have to check so many pixels
+    private RenderTexture blitTexture;
+
+    // Texture2D to copy the blitted texture to for CPU checking stuff
+    private Texture2D blitTexture2D;
+
+    // Layer for camera detection to render
+    int layerMask;
+
+    // All the fields below are for debugging
+    private GameObject[] debugPlanes;
+    public Material greenMat;
+    public Material redMat;
+
     // Start is called before the first frame update
     void Start()
     {
+        layerMask = 0 | (1 << LayerMask.NameToLayer("RedCube"));
+        Debug.Log(layerMask);
+        Camera.main.backgroundColor = Color.black;
+
         activePhotos = new List<Photo>();
         inactivePhotos = new Stack<Photo>();
         photosPlanes = new GameObject[MAX_PHOTOS];
+        debugPlanes = new GameObject[MAX_PHOTOS];
 
+        inspectTexture = new RenderTexture(512, 512, 16, RenderTextureFormat.ARGB32);
+        blitTexture = new RenderTexture(32, 32, 16, RenderTextureFormat.ARGB32);
+        blitTexture2D = new Texture2D(32, 32, TextureFormat.ARGB32, false);
 
         for (int i = 0; i < MAX_PHOTOS; i++)
         {
@@ -51,6 +82,10 @@ public class PhotoScript : MonoBehaviour
             planePos.x = (i - (MAX_PHOTOS / 2)) * 2;
             GameObject photoPlane = Instantiate(planePrefab, planePos, Quaternion.Euler(90, -90, 90));
             photosPlanes[i] = photoPlane;
+
+            planePos.y += 2;
+            GameObject debugPlane = Instantiate(planePrefab, planePos, Quaternion.Euler(90, -90, 90));
+            debugPlanes[i] = debugPlane;
 
             Material planeMaterial = new Material(Shader.Find("Standard"));
             photoPlane.GetComponent<MeshRenderer>().material = planeMaterial;
@@ -86,9 +121,44 @@ public class PhotoScript : MonoBehaviour
                     RenderTexture prevTarget = camera.targetTexture;
                     camera.targetTexture = curPhoto.texture;
                     camera.Render();
-                    camera.targetTexture = prevTarget;
 
                     activePhotos.Add(curPhoto);
+
+                    // Do the checking for our various flags
+                    int oldMask = camera.cullingMask;
+                    CameraClearFlags oldClearFlags = camera.clearFlags;
+                    camera.clearFlags = CameraClearFlags.SolidColor;
+                    camera.cullingMask = layerMask;
+                    camera.targetTexture = inspectTexture;
+                    camera.Render();
+
+                    // Cleanup
+                    camera.clearFlags = oldClearFlags;
+                    camera.cullingMask = oldMask;
+                    camera.targetTexture = prevTarget;
+
+                    // Check the pixels for non-black
+                    Graphics.Blit(inspectTexture, blitTexture);
+                    RenderTexture.active = blitTexture;
+                    blitTexture2D.ReadPixels(new Rect(0, 0, blitTexture.width, blitTexture.height), 0, 0);
+                    blitTexture2D.Apply();
+
+                    // Count non-black pixels
+                    int nonBlackPixels = 0;
+                    Color[] pixels = blitTexture2D.GetPixels();
+                    foreach (Color pixel in pixels)
+                    {
+                        if (pixel != Color.black)
+                        {
+                            nonBlackPixels++;
+                        }
+                    }
+
+                    // Check against threshold
+                    curPhoto.isRedCubeIn = (nonBlackPixels > REQ_AREA * blitTexture.width * blitTexture.height);
+
+                    // Cleanup
+                    RenderTexture.active = null; // Reset active RenderTexture
                 }
                 else
                 {
@@ -97,7 +167,7 @@ public class PhotoScript : MonoBehaviour
 
             }
         }
-        if (Input.GetKeyDown(KeyCode.R))
+        if (DEBUG_MODE && Input.GetKeyDown(KeyCode.R))
         {
             DeletePhoto(0);
         }
@@ -106,10 +176,16 @@ public class PhotoScript : MonoBehaviour
         for (int i = 0; i < MAX_PHOTOS; i++)
         {
             GameObject curPlane = photosPlanes[i];
+            GameObject curDebugPlane = debugPlanes[i];
             Material mat = curPlane.GetComponent<MeshRenderer>().material;
+            curDebugPlane.GetComponent<MeshRenderer>().material = redMat;
             if (activePhotos.Count > i)
             {
                 mat.mainTexture = activePhotos[i].texture;
+                if (activePhotos[i].isRedCubeIn)
+                {
+                    curDebugPlane.GetComponent<MeshRenderer>().material = greenMat;
+                }
             }
             else
             {
@@ -127,6 +203,16 @@ public class PhotoScript : MonoBehaviour
                 rt.Release();
             }
         }
+
+        if (inspectTexture && inspectTexture.IsCreated())
+        {
+            inspectTexture.Release();
+        }
+        if (blitTexture && blitTexture.IsCreated())
+        {
+            blitTexture.Release();
+        }
+
     }
 
     /** Delete the ith photo (0-indexed) */
